pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrobbledb"
version = "1.0.0"
description = "Save data from last.fm/libre.fm to a SQLite database"
authors = [
    {name = "Jacob Kaplan-Moss", email = "jacob@jacobian.org"},
    {name = "Brian M. Dennis"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "pylast>=3.1.0",
    "sqlite-utils>=1.12.1",
    "click>=7.0",
    "python-dateutil>=2.8.1",
    "platformdirs>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=5.2.2",
    "pytest-cov>=2.8.1",
]

[project.scripts]
scrobbledb = "scrobbledb.cli:cli"

[project.urls]
Repository = "https://github.com/jacobian/lastfm-to-sqlite/"

[dependency-groups]
dev = [
    "pytest>=7.0.1",
    "pytest-cov>=4.0.0",
]
